+++ src/Rsse.Service/Api/Controllers/SystemController.cs
@@ -1,6 +1,8 @@
+using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
 using SearchEngine.Domain.Configuration;
+using SearchEngine.Domain.Contracts;
 
 namespace SearchEngine.Api.Controllers;
 
@@ -8,12 +10,11 @@ namespace SearchEngine.Api.Controllers;
 /// Контроллер, поставляющий системную информацию.
 /// </summary>
 [ApiController]
-public class SystemController(IOptionsSnapshot<DatabaseOptions> options) : ControllerBase
+public class SystemController(IOptionsSnapshot<DatabaseOptions> options, ITokenizerService tokenizer) : ControllerBase
 {
     /// <summary>
     /// Получить версию сервиса.
     /// </summary>
-    /// <returns></returns>
     [HttpGet(RouteConstants.SystemVersionGetUrl)]
     public ActionResult GetVersion()
     {
@@ -24,5 +25,15 @@ public class SystemController(IOptionsSnapshot<DatabaseOptions> options) : Contr
             ReaderContext = options.Value.ReaderContext,
             CreateTablesOnPgMigration = options.Value.CreateTablesOnPgMigration
         });
+    }
+
+    /// <summary>
+    /// Дождаться прогрева токенизатора
+    /// </summary>
+    [HttpGet(RouteConstants.SystemWaitWarmUpGetUrl)]
+    public async Task<ActionResult> WaitReadiness()
+    {
+        await tokenizer.WaitWarmUp();
+        return Ok();
     }
 }